Move all PC Schemes to one module; bench-templates as a separate module
