+++ .github/workflows/ci.yml
@@ -96,11 +96,13 @@ jobs:
         with:
           use-cross: true
           command: build
-          args: --no-default-features --target aarch64-unknown-none
+          args: --workspace --no-default-features --target aarch64-unknown-none --
+exclude ark-pcs-bench-templates
 
       - name: Check
         uses: actions-rs/cargo@v1
         with:
           use-cross: true
           command: check
-          args: --examples --no-default-features --target aarch64-unknown-none
+          args: --workspace --examples --no-default-features --target aarch64-unknown-none --
+exclude ark-pcs-bench-templates
