+++ Cargo.toml
@@ -1,49 +1,19 @@
-[package]
-name = "ark-poly-commit"
+[workspace]
+members = ["pcs", "bench-templates"]
+resolver = "2"
+
+
+[workspace.package]
 version = "0.4.0"
+authors = ["arkworks contributors"]
 description = "A library for constructing polynomial commitment schemes for use in zkSNARKs"
 repository = "https://github.com/arkworks-rs/poly-commit"
 documentation = "https://docs.rs/ark-poly-commit/"
 keywords = ["cryptography", "commitments", "elliptic-curves", "pairing"]
-categories = ["cryptography"]
 include = ["Cargo.toml", "src", "README.md", "LICENSE-APACHE", "LICENSE-MIT"]
+categories = ["cryptography"]
 license = "MIT/Apache-2.0"
 edition = "2018"
-
-[dependencies]
-ark-serialize = { version = "^0.4.0", default-features = false, features = [ "derive" ] }
-ark-ff = { version = "^0.4.0", default-features = false }
-ark-ec = { version = "^0.4.0", default-features = false }
-ark-poly = {version = "^0.4.0", default-features = false }
-ark-crypto-primitives = {version = "^0.4.0", default-features = false, features = ["sponge", "merkle_tree"] }
-ark-std = { version = "^0.4.0", default-features = false }
-
-ark-relations = { version = "^0.4.0", default-features = false, optional = true }
-ark-r1cs-std = { version = "^0.4.0", default-features = false, optional = true }
-hashbrown = { version = "0.13", default-features = false, optional = true }
-
-digest = "0.10"
-derivative = { version = "2", features = [ "use_core" ] }
-rayon = { version = "1", optional = true }
-criterion = { version = "0.5", default-features = false, optional = true }
-rand_chacha = { version = "0.3.0", default-features = false, optional = true }
-
-[dev-dependencies]
-ark-ed-on-bls12-381 = { version = "^0.4.0", default-features = false }
-ark-bls12-381 = { version = "^0.4.0", default-features = false, features = [ "curve" ] }
-ark-bls12-377 = { version = "^0.4.0", default-features = false, features = [ "curve" ] }
-blake2 = { version = "0.10", default-features = false }
-rand_chacha = { version = "0.3.0", default-features = false }
-
-[[bench]]
-name = "pcs"
-path = "benches/benches.rs"
-harness = false
-
-[[bench]]
-name = "pcs-size"
-path = "benches/benches_size.rs"
-harness = false
 
 [profile.release]
 opt-level = 3
@@ -56,14 +26,6 @@ opt-level = 3
 debug-assertions = true
 incremental = true
 debug = true
-
-[features]
-default = [ "std", "parallel" ]
-std = [ "ark-ff/std", "ark-ec/std", "ark-poly/std", "ark-std/std", "ark-relations/std", "ark-serialize/std", "ark-crypto-primitives/std"]
-r1cs = [ "ark-relations", "ark-r1cs-std", "hashbrown", "ark-crypto-primitives/r1cs"]
-print-trace = [ "ark-std/print-trace" ]
-parallel = [ "std", "ark-ff/parallel", "ark-ec/parallel", "ark-poly/parallel", "ark-std/parallel", "rayon" ]
-benches = [ "default", "criterion", "rand_chacha" ]
 
 [patch.crates-io]
 ark-ff = { git = "https://github.com/arkworks-rs/algebra/" }
