+++ pcs/src/challenge.rs
