+++ pcs/src/error.rs
