+++ pcs/src/optional_rng.rs
